upate

+++ back-end/back-end/Service/Implement/PostService.cs
@@ -1,14 +1,15 @@
 ﻿using AutoMapper;
+using back_end.Common;
 using back_end.DTO;
 using back_end.DTO.UserDTOModel;
 using back_end.Enity;
 using back_end.Enum;
-using back_end.Models.Request;
 using back_end.Models.Response;
 using back_end.Respositories.Implement;
 using back_end.Respositories.Interface;
 using back_end.Service.Interface;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace back_end.Service.Implement
 {
@@ -18,6 +19,9 @@ namespace back_end.Service.Implement
         private IPostTechnologyRespository _postTechnologyRespository;
         private ITechnologyRespository _technologyRespository;
         private IEmployeeRespository _employeeRespository;
+        private IHttpContextAccessor _contextAccessor;
+        private IUserRespository _userRespository;
+        private IUserService _userService;
         private IMapper _mapper;
 
         public PostService(
@@ -25,12 +29,18 @@ namespace back_end.Service.Implement
             IPostTechnologyRespository postTechnologyRespository,
             ITechnologyRespository technologyRespository,  
             IEmployeeRespository employeeRespository,
+            IHttpContextAccessor contextAccessor,
+            IUserRespository userRespository,
+            IUserService userService,
             IMapper mapper)
         {
             _postRespository = postRespository;
             _postTechnologyRespository = postTechnologyRespository;
             _technologyRespository = technologyRespository;
             _employeeRespository = employeeRespository;
+            _contextAccessor = contextAccessor;
+            _userRespository = userRespository;
+            _userService = userService;
             _mapper = mapper;
         }
 
@@ -39,43 +49,46 @@ namespace back_end.Service.Implement
             var result = new AppResponse<PostDTO>();
             try
             {
+
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Can't found user from token");
+
                 // ------------ check employee and  permission -------------- //
                 var employee = await _employeeRespository
-                    .FindBy(em => !em.IsDelete && em.Id == post.EmployeeId)
-                    .Include(em=>em.User)
-                    .ThenInclude(u => u.UserRoles)
-                    .ThenInclude(ur=>ur.Role)
+                    .FindBy(em => !em.IsDelete && em.UserId == user.Id)
                     .FirstOrDefaultAsync();
-                if(
-                    employee is null 
-                    || employee.User.UserRoles.Where(ur => ur.Role.Name.Equals("business")).FirstOrDefault() is null
-                    || employee.Type != EmployeeType.OfficalEmployee 
-                ) return result.BuilderError("You don't have permission");
+                if( employee is null || employee.Type != EmployeeType.OfficalEmployee) 
+                    return result.BuilderError("You don't have permission");
 
 
                 // -------- saving new employee ---------------//
                 Post newPost = _mapper.Map<Post>(post);
                 newPost.InitialEnity();
+                newPost.EmployeeId = employee.Id;
                 await _postRespository.Insert(newPost);
 
                 // -------- saving post of technologies ------- //
                 List<PostTechnology> postTechnologyList = new List<PostTechnology>();
-                var techList = post.Technologies.Select(t => _mapper.Map<Technology>(t)).ToList();
+                var techList = post.Technologies?.Select(t => _mapper.Map<Technology>(t)).ToList();
                 post.Id = newPost.Id;
                 post.Technologies = [];
-                foreach (var technology in techList)
+                if(techList != null)
                 {
-                    var tech = await _technologyRespository.FindBy(t => t.Id == technology.Id && !t.IsDelete).FirstOrDefaultAsync();
-                    if (tech != null){
-                        var postTechnology = new PostTechnology();
-                        postTechnology.InitialEnity();
-                        postTechnology.PostId = newPost.Id;
-                        postTechnology.TechnologyId = tech.Id;
-                        post.Technologies.Add(_mapper.Map<TechnologyDTO>(tech));
-                        postTechnologyList.Add(postTechnology);
+                    foreach (var technology in techList)
+                    {
+                        var tech = await _technologyRespository.FindBy(t => t.Id == technology.Id && !t.IsDelete).FirstOrDefaultAsync();
+                        if (tech != null){
+                            var postTechnology = new PostTechnology();
+                            postTechnology.InitialEnity();
+                            postTechnology.PostId = newPost.Id;
+                            postTechnology.TechnologyId = tech.Id;
+                            post.Technologies.Add(_mapper.Map<TechnologyDTO>(tech));
+                            postTechnologyList.Add(postTechnology);
+                        }
                     }
+                    await _postTechnologyRespository.Insert(postTechnologyList);
                 }
-                await _postTechnologyRespository.Insert(postTechnologyList);
 
                 return result.BuilderResult(post,"Success");
             }
@@ -91,9 +104,15 @@ namespace back_end.Service.Implement
             var result = new AppResponse<bool>();
             try
             {
-                var post = await _postRespository.FindBy(p => p.Id == postId).FirstOrDefaultAsync();
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Not found user from token");
+
+                var post = await _postRespository.FindBy(p => p.Id == postId).Include(p=>p.Employee).FirstOrDefaultAsync();
                 if (post is null) return result.BuilderError("Not found post");
 
+                if (post.Employee.UserId != user.Id && !_userService.checkRole())
+                    return result.BuilderError("You are not Author of this post");
                 post.DeleteEnity();
                 await _postRespository.Update(post);
 
@@ -117,9 +136,8 @@ namespace back_end.Service.Implement
                         Id = p.Id,
                         Name = p.Name,
                         Context = p.Context,
-                        EmployeeId = p.EmployeeId,
                         ExperienceYear = p.ExperienceYear,
-                        Exprised = p.Exprised.ToString("dd/MM/yyyy"),
+                        Exprised = p.Exprised,
                         Technologies = p.PostTechnologies
                         .Where(pt => !pt.IsDelete && pt.PostId == p.Id)
                         .Select(pt => _mapper.Map<TechnologyDTO>(pt.Technology)).ToList(),
@@ -141,30 +159,32 @@ namespace back_end.Service.Implement
             }
         }
 
-        public Task<AppResponse<SearchResponse<PostDTO>>> Search(SearchResquest resquest)
-        {
-            throw new NotImplementedException();
-        }
-
         public async Task<AppResponse<PostDTO>> Update(PostDTO post)
         {
             var result = new AppResponse<PostDTO>();
             try
             {
-                var data = await _postRespository
+                Post? data = await _postRespository
                     .FindBy(p => !p.IsDelete && p.Id == post.Id)
                     .Include(p => p.Employee)
                     .Include(p=>p.PostTechnologies).ThenInclude(pt => pt.Technology)
                     .FirstOrDefaultAsync();
                 if (data is null) return result.BuilderError("Not found post");
 
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Can't found user from token");
+
+                if (user.Id != data.Employee.UserId && !_userService.checkRole())
+                    return result.BuilderError("You are not Author of this Post");
+
                 // ----------- Update information post ------------------- //
 
                 data.Context = post.Context;
                 data.Name = post.Name;
                 data.ExperienceYear = post.ExperienceYear;
-                data.Exprised = DateTime.Parse(post.Exprised);
-                data.UpdateTimeEntity();
+                data.Exprised = post.Exprised;
+
                 await _postRespository.Update(data);
 
                 // ---------- Get all technology id of post --------------- //
@@ -172,13 +192,13 @@ namespace back_end.Service.Implement
                 var newTechPostList = post.Technologies.Select(t => _mapper.Map<Technology>(t)).Select(t=>t.Id);
                 var addPostTechnologyList = new List<PostTechnology>();
                 var deletePostTechnologyList = new List<PostTechnology>();
-                PostTechnology? postTechnology;
+
                 // ----------- Add new technology for post --------------- //
-                foreach (var tectId in newTechPostList)
+                foreach(var tectId in newTechPostList)
                 {
                     if (!oldTechPostList.Contains(tectId))
                     {
-                        postTechnology = new PostTechnology();
+                        PostTechnology postTechnology = new PostTechnology();
                         postTechnology.InitialEnity();
                         postTechnology.PostId = data.Id;
                         postTechnology.TechnologyId = tectId;
@@ -187,16 +207,15 @@ namespace back_end.Service.Implement
                 }
                 await _postTechnologyRespository.Insert(addPostTechnologyList);
                 // --------------- Delete technology for post ------------ //
-                
-                foreach (var techId in oldTechPostList)
+                foreach(var techId in oldTechPostList)
                 {
                     if (!newTechPostList.Contains(techId))
                     {
-                        postTechnology = await _postTechnologyRespository
-                        .FindBy(pt => !pt.IsDelete && pt.TechnologyId == techId && pt.PostId == data.Id)
-                        .FirstOrDefaultAsync();
+                        PostTechnology? postTechnology = await _postTechnologyRespository
+                             .FindBy(pt => !pt.IsDelete && pt.TechnologyId == techId && pt.PostId == data.Id)
+                             .FirstOrDefaultAsync();
 
-                        if (postTechnology != null)
+                        if (postTechnology != null) 
                             deletePostTechnologyList.Add(postTechnology);
                     }
                 }
